==> R/00_all.qmd <==
---
title: "00_all"
format: 
  html:
    embed-resources: true
editor: visual
---

# Full project

This is joined quarto document, that runs all project steps.

```{r child = "01_load.qmd"}
```

```{r child = "02_clean.qmd"}
```

```{r child = "03_augment.qmd"}
```

```{r child = "04_describe.qmd"}
```

```{r child = "05_analysis_1.qmd"}
```

```{r child = "06_analysis_2.qmd"}
```

```{r child = "07_analysis_3.qmd"}
```
